requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.50.0
fpdf==1.7.2
idna==3.6
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.6
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qualysapi==8.1.0
requests==2.31.0
secure-smtplib==0.1.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
